rework acceptors
